Fix bug #66387: Stack overflow with imagefilltoborder
Assigned CVE-IDs: CVE-2015-8874

The stack overflow is caused by the recursive algorithm in combination with a
very large negative coordinate passed to gdImageFillToBorder(). As there is
already a clipping for large positive coordinates to the width and height of
the image, it seems to be consequent to clip to zero also.

(cherry picked from commit f96ebb098697908641c6d37e1149cde1f90417fd)
Signed-off-by: Lior Kaplan <kaplanlior@gmail.com>